Split services into different crates
