+++ Cargo.toml
@@ -1,5 +1,7 @@
 [workspace]
 
 members = [
-    "nomos-services"
+    "nomos-services/log",
+    "nomos-services/network",
+    "nomos-services/storage"
 ]
