+++ nomos-services/log/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "nomos-log"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = "0.1"
+overwatch-rs = { git = "https://github.com/logos-co/Overwatch", branch = "main" }
+serde = "1.0"
+tracing = "0.1"
+tracing-appender = "0.2"
+tracing-subscriber =  { version = "0.3", features = ["json"] }
+tracing-gelf = "0.7"
+futures = "0.3"
