+++ nomos-services/network/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "nomos-services"
+name = "nomos-network"
 version = "0.1.0"
 edition = "2021"
 
@@ -15,17 +15,11 @@ sled = { version = "0.34", optional = true }
 tokio = { version = "1", features = ["sync"] }
 thiserror = "1.0"
 tracing = "0.1"
-waku = { git = "https://github.com/waku-org/waku-rust-bindings" }
+waku = { git = "https://github.com/waku-org/waku-rust-bindings", optional = true }
 tracing-appender = "0.2"
 tracing-subscriber =  { version = "0.3", features = ["json"] }
 tracing-gelf = "0.7"
 futures = "0.3"
 
-[dev-dependencies]
-tempfile = "3.3"
-tokio = { version = "1", features = ["sync", "macros", "time"] }
-
 [features]
-default = []
-mock = []
-sled-backend = ["sled"]
+default = ["waku"]
