+++ nomos-services/network/src/backends/mod.rs
@@ -2,7 +2,9 @@ use super::*;
 use overwatch_rs::services::state::ServiceState;
 use tokio::sync::broadcast::Receiver;
 
+#[cfg(feature = "waku")]
 mod waku;
+#[cfg(feature = "waku")]
 pub use self::waku::Waku;
 
 #[async_trait::async_trait]
