--- nomos-services/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod log;
-pub mod network;
-pub mod storage;
